Split up health check endpoints and testing endpoints
