+++ breadbox/breadbox/api/__init__.py
@@ -14,7 +14,8 @@ from .compute import router as compute_router
 from .user import router as user_router
 from .metadata import router as metadata_router
 from .temp import router as temp_router
-from breadbox.health_check.health_check import router as health_check_router
+from .dev.health_check import router as health_check_router
+from .dev.dev_test import router as dev_test_router
 from breadbox.schemas.custom_http_exception import ERROR_RESPONSES
 
 api_router = APIRouter(responses=ERROR_RESPONSES)  # type: ignore
@@ -31,4 +32,5 @@ api_router.include_router(partials_router)
 api_router.include_router(user_router)
 api_router.include_router(metadata_router)
 api_router.include_router(health_check_router)
+api_router.include_router(dev_test_router)
 api_router.include_router(temp_router)
